kaspi_bot/requirements.txt
aiogram==3.15.0
httpx==0.28.1
apscheduler==3.11.0
pypdf==5.1.0
python-dotenv==1.0.1
reportlab==4.0.4
